MedicalDialogueTaskGenerator/requirements.txt
# 核心依赖
pyyaml>=6.0
jsonschema>=4.0.0
pydantic>=2.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 文本处理
jieba>=0.42.0
transformers>=4.30.0

# 工具库
tqdm>=4.65.0
click>=8.1.0
python-dotenv>=1.0.0

# 日志
colorlog>=6.7.0

# 类型检查
typing-extensions>=4.5.0

# HTTP请求（可选，用于在线医学知识库）
requests>=2.31.0

# 测试（开发依赖）
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
